moved all relevant files to tools.attn_maps_plotting
