+++ tools/attn_maps_plotting/attn_scores_plotting.py
@@ -10,7 +10,7 @@ import os
 from typing import List, Tuple, Iterable, Dict
 import torch
 from src.datasets.dataset_config import DatasetConfig
-from tools.plot_attn_score_config import PlotAttnScoreConfig
+from tools.attn_maps_plotting.plot_attn_score_config import PlotAttnScoreConfig
 from src.datasets.label_utils import get_unique_parts_from_labels, get_markers_from_labels, get_batches_from_labels
 
 def get_percentiles(data, prc_list = [25,50,75], axis=0):
