+++ tools/attn_maps_plotting/plot_attn_score_config.py
